Rename tools to output
