+++ input/cpd.go
@@ -4,10 +4,10 @@ import (
 	"encoding/xml"
 	"fmt"
 
-	"github.com/mallowlabs/scarfco/tools"
+	"github.com/mallowlabs/scarfco/output"
 )
 
-func ConvertCPD(content []byte) *tools.Result {
+func ConvertCPD(content []byte) *output.Result {
 	type File struct {
 		Line int    `xml:"line,attr"`
 		Path string `xml:"path,attr"`
@@ -52,7 +52,7 @@ func ConvertCPD(content []byte) *tools.Result {
 		}
 	}
 
-	result := tools.Result{}
+	result := output.Result{}
 	for k, v := range m {
 		file := result.AddFile(k)
 		for _, rf := range v {
