+++ input/pmd.go
@@ -4,10 +4,10 @@ import (
 	"encoding/xml"
 	"strings"
 
-	"github.com/mallowlabs/scarfco/tools"
+	"github.com/mallowlabs/scarfco/output"
 )
 
-func ConvertPMD(content []byte) *tools.Result {
+func ConvertPMD(content []byte) *output.Result {
 	type Violation struct {
 		Beginline int    `xml:"beginline,attr"`
 		Priority  int    `xml:"priority,attr"`
@@ -29,7 +29,7 @@ func ConvertPMD(content []byte) *tools.Result {
 	var pmd Pmd
 	xml.Unmarshal(content, &pmd)
 
-	result := tools.Result{}
+	result := output.Result{}
 	for _, file := range pmd.Files {
 		f := result.AddFile(file.Name)
 		for _, violation := range file.Violations {
